chore: move bench script
